Fix race condition with APIRequesterWrapper
The initial call to GetAPIRequester (when apiRequesterWrapper.apiRequester
is nil) is now thread safe. Furthermore, changed the mutex protecting
the wrapper to a pointer. It's my understanding that mutex' aren't
safe to copy.

(Running the integration tests with -race confirmed this was an issue,
so this was added to the README)